bugs fixed

+++ Unicepse/Commands/Payments/DeletePaymentCommand.cs
@@ -25,19 +25,24 @@ namespace Unicepse.Commands.Payments
         }
         public override async Task ExecuteAsync(object? parameter)
         {
-            _subscriptionDataStore.SelectedSubscription = _paymentDataStore.SelectedPayment!.Subscription;
-            if (_subscriptionDataStore.SelectedSubscription!.Player == null)
+            if (MessageBox.Show("سيتم حذف هذا الدفعة , هل انت متاكد", "تنبيه", MessageBoxButton.YesNo,
+                                          MessageBoxImage.Warning) == MessageBoxResult.Yes)
             {
-                _subscriptionDataStore.SelectedSubscription!.Player = new Core.Models.Player.Player() { Id = _playersDataStore.SelectedPlayer!.Id };
+                _subscriptionDataStore.SelectedSubscription = _paymentDataStore.SelectedPayment!.Subscription;
+                if (_subscriptionDataStore.SelectedSubscription!.Player == null)
+                {
+                    _subscriptionDataStore.SelectedSubscription!.Player = new Core.Models.Player.Player() { Id = _playersDataStore.SelectedPlayer!.Id };
+                }
+                _subscriptionDataStore.SelectedSubscription!.IsPaid = false;
+                _subscriptionDataStore.SelectedSubscription.PaidValue -= _paymentDataStore.SelectedPayment!.PaymentValue;
+                _playersDataStore.SelectedPlayer!.Player.Balance -= _paymentDataStore.SelectedPayment!.PaymentValue;
+
+                await _playersDataStore.UpdatePlayer(_playersDataStore.SelectedPlayer!.Player);
+                await _paymentDataStore.Delete(_paymentDataStore.SelectedPayment!);
+                await _subscriptionDataStore.Update(_subscriptionDataStore.SelectedSubscription!);
+                MessageBox.Show("تم حذف الدفعة بنجاح");
             }
-            _subscriptionDataStore.SelectedSubscription!.IsPaid = false;
-            _subscriptionDataStore.SelectedSubscription.PaidValue -= _paymentDataStore.SelectedPayment!.PaymentValue;
-            _playersDataStore.SelectedPlayer!.Player.Balance -= _paymentDataStore.SelectedPayment!.PaymentValue;
-            
-            await _playersDataStore.UpdatePlayer(_playersDataStore.SelectedPlayer!.Player);
-            await _paymentDataStore.Delete(_paymentDataStore.SelectedPayment!);
-            await _subscriptionDataStore.Update(_subscriptionDataStore.SelectedSubscription!);
-            MessageBox.Show("تم حذف الدفعة بنجاح");
+              
         }
     }
 }
